backend/pyproject.toml
[project]
name = "novaleap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.135.3",
    "httpx>=0.27.0",
    "pgvector>=0.3.6",
    "psycopg2>=2.9.11",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sentence-transformers>=3.4.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]
